+++ src/Home.jsx
@@ -115,33 +115,7 @@ function Member({ giho, name, jd_name, age }) {
 // - 5번 : 시·도의회의원선거 // 서울특별시 마포구 까지 요청, SGG_NAME 추가로 표기(마포구제1선거구)
 // - 11번 : 교육감선거   // 서울특별시 까지 요청
 
-const getLocation = () => {
-  let lat, long;
-  if (navigator.geolocation) {
-    // GPS를 지원하면
-    navigator.geolocation.getCurrentPosition(
-      function (position) {
-        lat = position.coords.latitude;
-        long = position.coords.longitude;
-        alert('위도 : ' + lat + ' 경도 : ' + long);
-      },
-      function (error) {
-        console.error(error);
-      },
-      {
-        enableHighAccuracy: false,
-        maximumAge: 0,
-        timeout: Infinity,
-      },
-    );
-  } else {
-    alert('GPS를 지원하지 않습니다');
-    return;
-  }
-};
-
 function Home({ candidates, num }) {
-  getLocation();
   let sg_type = [
     null,
     null,
