+++ src/hooks/useCandiate.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import useAPI from './useAPI';
 import getURL from './getURL';
+import run from './geocoding';
 // - 2번: 국회의원 선거
 // - 3번: 시 도지사 선거  // 서울특별시 까지 요청
 // - 4번 : 구,시,군의 장 선거 // 서울특별시 마포구 까지 요청
@@ -13,7 +14,9 @@ function useCandidate(num) {
 
   useEffect(() => {
     const fetchCandidate = async () => {
-      const URL = getURL(num);
+      const address = await run();
+      console.log(address);
+      const URL = getURL(num, address[1], address[2]);
       const json = await customFetch(URL);
       setCandidates([...json.getPofelcddRegistSttusInfoInqire.item]); // 4번일 때
     };
